requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Django==5.1.7
django-extensions==4.1
django-filter==25.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
Flask==3.1.0
git-filter-repo==2.47.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2

mysqlclient==2.2.7
pycparser==2.22

python-decouple==3.8
requests==2.32.3
SQLAlchemy==2.0.39
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise
pandas
openpyxl